del 15 al 29

+++ tema5/ejer15/src/ejer15.java
@@ -5,10 +5,10 @@ public class ejer15 {
   public static void main(String[] args) throws Exception {
     Scanner s = new Scanner(System.in);
     System.out.print("Introduzca el numero base: ");
-    double base = s.nextDouble();
+    int base = s.nextInt();
     System.out.print("Introduzca un numero exponente: ");
     int exponenteFinal = s.nextInt();
-    double potencia;
+    int potencia;
     int exponente;
     for (int i = 1; i <= exponenteFinal; i++) {
       potencia = 1;
